requirements.txt
google-genai
chromadb
python-dotenv
pyTelegramBotAPI